delete_sync/requirements.txt
azure-functions==1.21.3
azure-storage-blob==12.23.1
python-dotenv==1.0.1
snowflake-connector-python==3.12.2
cryptography==43.0.1

